+++ NEWS.md
@@ -7,12 +7,14 @@ Changes:
 
 2. Switched Oracle translation of `SELECT TOP` from `WHERE ROWNUM <=` to `FETCH FIRST ROWS ONLY`.
 
+
 Bugfixes:
 
 1. Fixed erroneously identifying parts of quoted text as start of SQL comments.
 
 2. Fixed translation of `INSERT INTO` with more than one CTE for Spark.
 
+3. Fixed translation of `SELECT TOP` on Oracle when using `DISTINCT`.
 
 SqlRender 1.10.0
 ================